requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Mail==0.9.1
Flask-Sitemap==0.4.0
Flask-xCaptcha==0.5.5
gunicorn==21.2.0
idna==3.3
importlib-metadata==6.8.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
myfuncs==1.4.8
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
sitemap==20191121
typing_extensions==4.8.0
urllib3==1.26.8
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
